Improvements to statemanager and also fixed some input
Changes to "State" where there is now a initialize function that will be used for all of the inits.
There is an reInitialize function which ONLY gets called IF a state that was "above" gets poped. Some stuff needs to get re-initialized when that happens. More info in the code ;)