Movement & camera changes
NOTE:
Camera is zoomed out by 2x atm.
Also made the whole thing framerate independent which it wasn't :alien:

Made some small changes to the camera, it now follows the player with "lerp" instead of being locked on all the time

Controller is now updated to what we discussed. The movement of the player might still need some tweeking but it will do for now